Moving the enum Currency to model package
