+++ src/main/java/ru/intervale/TeamProject/controller/BookPrice.java
@@ -13,7 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import ru.intervale.TeamProject.model.book.Currency;
+import ru.intervale.TeamProject.model.Currency;
 
 import javax.validation.Valid;
 public interface BookPrice {
